+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## HEAD (unreleased)
 
+- Enhancement: Add Dropzone component
+- Fix(FileButtonComponent): change id to be unique
+
+=======
+
 ## 35.1.2 (2021-03-15)
 
 - Re-publish master on npm `latest` channel. No changes